+++ packages/vite/src/node/logger.ts
@@ -26,10 +26,6 @@ export interface LogErrorOptions extends LogOptions {
   error?: Error | RollupError | null
 }
 
-export const buildFailInfo: Record<string, boolean> = {
-  failed: false,
-}
-
 export const LogLevels: Record<LogLevel, number> = {
   silent: 0,
   error: 1,
